+++ Realm/Realm/Linq/RealmResults.cs
@@ -16,6 +16,8 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
+using Realms.Helpers;
+using Realms.Schema;
 using System;
 using System.Linq;
 using System.Linq.Expressions;
@@ -70,9 +72,23 @@ namespace Realms
 
         #region IList members
 
-        public override bool Contains(object value) => Contains((T)value);
+        public override int IndexOf(T value)
+        {
+            Argument.NotNull(value, nameof(value));
 
-        public override int IndexOf(object value) => IndexOf((T)value);
+            if (_argumentType != (PropertyType.Object | PropertyType.Nullable))
+            {
+                throw new NotSupportedException("IndexOf on non-object results is not supported.");
+            }
+
+            var obj = Operator.Convert<T, RealmObject>(value);
+            if (!obj.IsManaged || obj.Realm.IsSameInstance(Realm))
+            {
+                throw new ArgumentException("Value does not belong to a realm", nameof(value));
+            }
+
+            return ResultsHandle.Find(obj.ObjectHandle);
+        }
 
         #endregion
     }