+++ wrappers/src/results_cs.cpp
@@ -170,4 +170,14 @@ REALM_EXPORT Results* results_snapshot(const Results& results, NativeException::
     });
 }
 
+REALM_EXPORT size_t results_find_object(Results& results, const Object& object_ptr, NativeException::Marshallable& ex)
+{
+	return handle_errors(ex, [&]() {
+		if (results.get_realm() != object_ptr.realm()) {
+			throw ObjectManagedByAnotherRealmException("Can't look up index of an object that belongs to a different Realm.");
+		}
+		return results.index_of(object_ptr.row());
+	});
+}
+
 }   // extern "C"